refactor: CartItem이 아닌 Product로 네이밍 수정
