+++ app/src/main/java/nextstep/shoppingcart/ui/component/Product.kt
@@ -15,12 +15,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
 import nextstep.shoppingcart.R
-import nextstep.shoppingcart.ui.screen.products.model.Product
+import nextstep.shoppingcart.ui.screen.products.model.ProductModel
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
 
 @Composable
-fun CartItem(
-    product: Product,
+fun Product(
+    productModel: ProductModel,
     modifier: Modifier = Modifier
 ) {
     Column(modifier = modifier) {
@@ -28,19 +28,19 @@ fun CartItem(
             modifier = Modifier
                 .width(156.dp)
                 .height(158.dp),
-            model = product.imageUrl,
-            contentDescription = stringResource(R.string.product_image_description, product.name),
+            model = productModel.imageUrl,
+            contentDescription = stringResource(R.string.product_image_description, productModel.name),
             placeholder = painterResource(id = R.drawable.ic_launcher_background)
         )
         Text(
-            text = product.name,
+            text = productModel.name,
             maxLines = 1,
             overflow = TextOverflow.Ellipsis,
             style = MaterialTheme.typography.titleMedium,
             fontWeight = FontWeight.Bold,
         )
         Text(
-            text = stringResource(R.string.product_price, product.price),
+            text = stringResource(R.string.product_price, productModel.price),
             overflow = TextOverflow.Ellipsis,
             style = MaterialTheme.typography.titleMedium,
         )
@@ -51,8 +51,8 @@ fun CartItem(
 @Composable
 private fun CartItemPreview() {
     ShoppingCartTheme {
-        CartItem(
-            Product(
+        Product(
+            ProductModel(
                 name = "iPhone 15 Pro Max",
                 imageUrl = "https://img.danawa.com/prod_img/500000/334/189/img/28189334_1.jpg",
                 price = 1_900_000
